Updated rule registration scheme
Every rule is now added to the model upon creation. The desired rules
may then be selected by the user.
We are now storing all rules in a dictionary. Active rules are in a list
selected from the dictionary. Checks are included to ensure, the same
rule is not selected twice.